requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
databricks-sdk>=0.49.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
anyio>=4.0.0
